requirements.txt
aiohttp
asyncio
discord.py
requests
discord
python-dotenv
PyYAML
youtube-transcript-api
PyPDF2
python-docx
tqdm
pyshorteners
yt_dlp
fake-useragent
pillow
openai
PyNaCl
psutil
cryptography
geopy
imagehash